remove console, event and queue explorer and unionize them into a single global explorer
