+++ lib/console/index.ts
@@ -3,7 +3,6 @@ export * from './decorators';
 export * from './logger';
 export * from './inquirer';
 export * from './commands';
-export * from './explorer';
 export * from './runner';
 export * from './interfaces';
 export * from './consoleIO';
