+++ lib/events/index.ts
@@ -1,7 +1,6 @@
 export * from './event';
 export * from './decorator';
 export * from './metadata';
-export * from './explorer';
 export * from './interfaces';
 export * from './helpers';
 export * from './jobListener';
